requirements.txt
prettytable==2.1.0
numpy==1.20.3
torch==1.8.1+cu101
torchvision==0.9.1+cu101
Pillow==9.1.1
scikit_learn==1.1.1
